// src/app/type/account.type.ts
export type TAccount = {
  last_login: string | null;
  username: string | null;
  first_name: string | null;
  last_name: string | null;
  date_joined: string | null;
  uuid: string | null;
  uid: string | null;
  role: string | null;
  address: string | null;
  address_legal: string | null;
  city: string | null;
  city_ccia_registration: string | null;
  company_certificate: string | null;
  company_name: string | null;
  fiscal_code: string | null;
  dob: string | null;
  email: string | null;
  nationality_code: string | null;
  num_rea: string | null;
  pec: string | null;
  phone_prefix: string | null;
  phone: string | null;
  state: string | null;
  recipient_code: string | null;
  updated_at: string | null;
  vat_number : string | null;
}